+++ src/frontend/js/widgets/Dashboard/components/DashboardLayoutRoute/index.tsx
@@ -1,0 +1,18 @@
+import { Outlet, useMatches } from 'react-router-dom';
+import { DashboardRouteHandle } from 'widgets/Dashboard/hooks/useDashboardRouter';
+import { DashboardLayout } from 'widgets/Dashboard/components/DashboardLayout';
+
+export const DashboardLayoutRoute = () => {
+  const matches = useMatches();
+  const renderOutletOnly = matches.find((match) => {
+    const handle = match.handle || {};
+    return !!(handle as DashboardRouteHandle).renderLayout;
+  });
+  return renderOutletOnly ? (
+    <Outlet />
+  ) : (
+    <DashboardLayout>
+      <Outlet />
+    </DashboardLayout>
+  );
+};