+++ src/frontend/js/widgets/Dashboard/components/DashboardSidebar/index.tsx
@@ -1,20 +1,15 @@
-import { defineMessages, FormattedMessage, useIntl } from 'react-intl';
+import { defineMessages, FormattedMessage, MessageDescriptor } from 'react-intl';
 import { matchPath, NavLink, useLocation, useNavigate } from 'react-router-dom';
 import { ChangeEvent, useMemo, useRef } from 'react';
 import { useSession } from 'contexts/SessionContext';
-import { DashboardPaths, getDashboardRouteLabel, getDashboardRoutePath } from '../../utils/routers';
-import { DashboardAvatar } from '../DashboardAvatar';
+import { DashboardAvatar } from 'widgets/Dashboard/components/DashboardAvatar';
+import SettingsLink from 'components/SettingsLink';
 
 const messages = defineMessages({
-  header: {
-    id: 'components.DashboardSidebar.header',
-    description: 'Title of the dashboard sidebar',
-    defaultMessage: 'Welcome {name}',
-  },
-  subHeader: {
-    id: 'components.DashboardSidebar.subHeader',
-    description: 'Sub title of the dashboard sidebar',
-    defaultMessage: 'You are on your dashboard',
+  settingsLinkLabel: {
+    id: 'components.DashboardSidebar.settingsLinkLabel',
+    description: 'label of the settings link',
+    defaultMessage: 'Settings',
   },
   responsiveNavLabel: {
     id: 'components.DashboardSidebar.responsiveNavLabel',
@@ -23,28 +18,27 @@ const messages = defineMessages({
   },
 });
 
-export const DashboardSidebar = () => {
+interface DashboardSidebarProps {
+  menuLinks: Record<string, string>[];
+  settingsUrl?: string;
+  header: MessageDescriptor;
+  subHeader: MessageDescriptor;
+}
+
+export const DashboardSidebar = ({
+  menuLinks,
+  settingsUrl,
+  header,
+  subHeader,
+}: DashboardSidebarProps) => {
   const { user } = useSession();
-  const intl = useIntl();
   const location = useLocation();
   const navigate = useNavigate();
   const selectNav = useRef<HTMLSelectElement>(null);
   const classes = ['dashboard__sidebar'];
 
-  const getRoutePath = getDashboardRoutePath(intl);
-  const getRouteLabel = getDashboardRouteLabel(intl);
-
-  const links = useMemo(
-    () =>
-      [DashboardPaths.COURSES, DashboardPaths.PREFERENCES].map((path) => ({
-        to: getRoutePath(path),
-        label: getRouteLabel(path),
-      })),
-    [],
-  );
-
   const selectedLink = useMemo(
-    () => links.find((link) => matchPath({ path: link.to, end: false }, location.pathname))?.to,
+    () => menuLinks.find((link) => matchPath({ path: link.to, end: false }, location.pathname))?.to,
     [location],
   );
 
@@ -61,11 +55,12 @@ export const DashboardSidebar = () => {
             <DashboardAvatar user={user!} />
           </div>
           <h3>
-            <FormattedMessage {...messages.header} values={{ name: user?.username }} />
+            <FormattedMessage {...header} values={{ name: user?.username }} />
           </h3>
           <p>
-            <FormattedMessage {...messages.subHeader} />
+            <FormattedMessage {...subHeader} />
           </p>
+          {settingsUrl && <SettingsLink to={settingsUrl} />}
         </header>
         <div className="dashboard__sidebar__container__responsive-nav">
           <label htmlFor="dashboard-responsive-nav" className="offscreen">
@@ -78,7 +73,7 @@ export const DashboardSidebar = () => {
             onChange={onSelectChange}
             className="form-field__select-input"
           >
-            {links.map((link) => (
+            {menuLinks.map((link) => (
               <option key={link.to} value={link.to}>
                 {link.label}
               </option>
@@ -87,7 +82,7 @@ export const DashboardSidebar = () => {
         </div>
 
         <ul>
-          {links.map((link) => (
+          {menuLinks.map((link) => (
             <li key={link.to}>
               <NavLink to={link.to}>{link.label}</NavLink>
             </li>